move NodeOwned
